Changes to the ai system now that we have hit the truth
