+++ src/main/kotlin/eater/ai/ashley/ConsideredAction.kt
@@ -4,13 +4,15 @@ import com.badlogic.ashley.core.Entity
 
 class ConsideredAction(
     name: String,
-    private val abortFunction: (entity: Entity) -> Unit,
-                       private val actFunction: (entity: Entity, deltaTime:Float) -> Unit,
-                       vararg consideration: Consideration
-): AiAction(name) {
+    scoreRange: ClosedFloatingPointRange<Float>,
+    private val actFunction: (entity: Entity, deltaTime: Float) -> Unit,
+    vararg consideration: Consideration
+) : AiAction(name,scoreRange) {
     init {
         considerations.addAll(consideration)
     }
+
+    var abortFunction: (entity: Entity) -> Unit = {}
     override fun abort(entity: Entity) {
         abortFunction(entity)
     }
